Extract query body excerpt partial
